seperated VR-related codes; fine tuned motion control of shield bash; adjust dominate muzzle of nemforcer when two handed
