+++ EnforcerMod_Unity/EnforcerVR/Assets/Prefabs/EnforcerShieldHand.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 68b279ebde483414085610904120ca02
+guid: 102ecd5fd7df74a4783cac6806dd05fa
 PrefabImporter:
   externalObjects: {}
   userData:
